+++ changelog.md
@@ -1,4 +1,7 @@
+CHANGES:
+- Added T3 production technologies
+- Added ROI for rune technologium
+
 FIXES:
 - Calculator settings button wordt nu niet meer dubbel aangemaakt.
 - De berekening om naar verzamelaar te gaan wordt nu ook goed uitgevoerd wanneer je meer processoren hebt dan het aantal dat gebruikt wordt.
-- Added T3 production technologies